Implemented a confirm dialog in bulk upload
