+++ src/bulk_upload/templates/bulk_upload_entry.html
@@ -5,13 +5,16 @@
 {% block head %}
 <link href="{% static 'css/upload.css' %}?v=1" rel="stylesheet">
 <link href="{% static 'css/multi-select.css' %}?v=1" rel="stylesheet">
+<link href="{% static 'css/jquery.dataTables.min.css' %}?v=1" rel="stylesheet">
 <link href="{% static 'css/dropzone.css' %}?v=1" rel="stylesheet">
 <link href="{% static 'css/feeds.css' %}?v=1" rel="stylesheet">
 <script src="{% static 'js/jquery.multi-select.js' %}?v=1"></script>
 <script src="{% static 'js/sns_common.js' %}?v=1"></script>
+<script src="{% static 'js/confirm_dialog.js' %}?v=1"></script>
 <script src="{% static 'js/upload.js' %}?v=1"></script>
 <script src="{% static 'js/jquery.quicksearch.js' %}?v=1"></script>
 <script src="{% static 'js/dropzone.js' %}?v=1"></script>
+<script src="{% static 'js/jquery.dataTables.min.js' %}?v=1"></script>
 {% endblock head %}
 
 {% block main %}
@@ -26,7 +29,7 @@
 
     <div class="row">
         <div class="col-xl-12 col-lg-12 col-md-12 col-sm-12 col-xs-12">
-            <form class="dropzone" name="upload-other-types" id="upload-form-other-types" action="{% url 'bulk_upload_post' %}" method="post" enctype="multipart/form-data">{% csrf_token %}</form>
+            <form class="dropzone" name="bulk_upload_post" id="bulk-upload-post" action="{% url 'bulk_upload_post' %}" method="post" enctype="multipart/form-data">{% csrf_token %}</form>
         </div>
     </div> <!-- <div class="row"> -->
  
@@ -37,13 +40,13 @@
             {% trans 'Title' %} :
         </div>
         <div class="colxl-4 col-lg-4 col-md-4 col-sm-4 col-xs-4">
-            <input type="text" id="stix-title-text" class="form-control" placeholder="{% trans '(Optional) Enter STIX Title ...' %}"/>
+            <input type="text" id="stix-title-text" class="form-control" placeholder="{% trans '(Optional) Enter STIX Title ...' %}" value="default" />
         </div><!--  <div class="col"> -->
         <div class="colxl-2 col-lg-2 col-md-2 col-sm-2 col-xs-2">
             {% trans 'Content' %} :
         </div>
         <div class="colxl-4 col-lg-4 col-md-4 col-sm-4 col-xs-4">
-            <input type="text" id="stix-description-text" class="form-control" placeholder="{% trans '(Optional) Enter STIX Description ...' %}"/>
+            <input type="text" id="stix-description-text" class="form-control" placeholder="{% trans '(Optional) Enter STIX Description ...' %}"/ value="default">
         </div><!--  <div class="col"> -->
     </div> <!-- <div class="row"> -->
 
@@ -54,7 +57,7 @@
             {% trans 'Merge into a single STIX file' %}:
         </div>
         <div class="colxl-4 col-lg-4 col-md-4 col-sm-4 col-xs-4">
-            <input type="checkbox" id="checkbox-is-single"  checked>
+            <input type="checkbox" id="checkbox-is-single" />
         </div>
         <div class="colxl-2 col-lg-2 col-md-2 col-sm-2 col-xs-2">
             {% trans 'TLP' %} :
@@ -87,7 +90,7 @@
             {% trans 'Display a confirm dialog' %}:
         </div>
         <div class="colxl-4 col-lg-4 col-md-4 col-sm-4 col-xs-4">
-            <input type="checkbox" id="checkbox-is-confirm-cee" disabled>
+            <input type="checkbox" id="checkbox-is-confirm-cee" checked/>
         </div>
         <div class="colxl-2 col-lg-2 col-md-2 col-sm-2 col-xs-2">
             {% trans 'Sharing Range' %} :
@@ -144,6 +147,23 @@
             <button type="button" id="upload-button" class="btn btn-lg btn-info">Upload</button>
         </div>
     </div> <!-- <div class="row"> -->
+
+    <div class="modal fade" id="confirm_indicators_modal_dialog" tabindex="-1">
+        <div class="modal-dialog modal-lg">
+            <div class="modal-content">
+                <div class="modal-header">
+                    <button type="button" class="close" data-dismiss="modal"><span></span></button>
+                    <h4 class="modal-title">Check Imported CTI Elements</h4>
+                </div>
+                <div class="modal-body" id='indicator-modal-body'>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-default" id="confirm-cancel" data-dismiss="modal">Cancel</button>
+                    <button type="button" class="btn btn-primary" id="confirm-compose">Compose</button>
+                </div>
+          </div>
+        </div>
+    </div>
 
 </div> <!-- <div class="container"> -->
 {% endblock main %}
