+++ src/feeds/templates/feeds/feeds.html
@@ -9,6 +9,7 @@
   <script src="{% static 'js/jquery.bullseye-1.0-min.js' %}"></script>
   <script src="{% static 'js/jquery.multi-select.js' %}?v=1"></script>
   <script src="{% static 'js/sns_common.js' %}?v=1"></script>
+  <script src="{% static 'js/confirm_dialog.js' %}?v=1"></script>
   <script src="{% static 'js/feeds.js' %}?v=1"></script>
   <script src="{% static 'js/jquery.quicksearch.js' %}?v=1"></script>
   <script src="{% static 'js/jquery.dataTables.min.js' %}?v=1"></script>
